pytransfer/requirements.txt
absl-py==0.4.0
alabaster==0.7.7
anaconda-client==1.4.0
anaconda-navigator==1.1.0
argcomplete==1.0.0
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
Automat==0.7.0
Babel==2.2.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
backports.weakref==1.0.post1
beautifulsoup4==4.4.1
bitarray==0.8.1
bleach==3.1.0
bokeh==1.0.4
boto==2.39.0
cdecimal==2.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
chest==0.2.3
cloudpickle==0.1.1
clyent==1.2.1
colorama==0.3.7
conda==4.5.12
conda-build==3.15.1
conda-manager==0.3.1
configobj==5.0.6
configparser==3.5.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
Cython==0.23.4
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.4
docopt==0.6.2
docutils==0.14
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
Flask==0.10.1
Flask-Cors==2.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gast==0.2.0
gevent==1.1.0
glob2==0.6
greenlet==0.4.9
grin==1.2.1
grpcio==1.14.1
HeapDict==1.0.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==4.1.1
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.10
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==4.1.1
jupyter-core==4.4.0
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
Keras==1.0.3
kiwisolver==1.0.1
llvmlite==0.9.0
locket==0.2.0
lockfile==0.12.2
luigi==2.7.2
lxml==3.6.0
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==2.2.3
memoize==1.0.0
memory-profiler==0.52.0
mistune==0.8.4
mock==2.0.0
mpmath==0.19
multipledispatch==0.4.8
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
networkx==1.11
nltk==3.2
nose==1.3.7
notebook==5.7.4
numpy==1.15.1
openpyxl==2.3.2
packaging==19.0
pandas==0.23.4
pandocfilters==1.4.2
param==1.8.1
parsel==1.5.0
path.py==0.0.0
pathlib2==2.3.3
patsy==0.5.0
pbr==4.2.0
pep8==1.7.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
pkginfo==1.5.0.1
ply==3.8
prometheus-client==0.5.0
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==4.1.0
ptyprocess==0.6.0
py==1.4.31
py-cpuinfo==4.0.0
py-d3==0.3.0
pyasn1==0.1.9
pyasn1-modules==0.2.2
pycairo==1.10.0
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0.2
PyDispatcher==2.0.5
pyflakes==1.1.0
Pygments==2.3.1
PyHamcrest==1.9.0
pymongo==3.7.1
pyOpenSSL==16.2.0
pyparsing==2.2.0
Pyro==3.16
pytest==2.8.5
python-daemon==2.1.2
python-dateutil==2.7.5
-e git+git@github.com:yusuke0519/TL_utils.git@bbfda70c48cf13d7afa85dd6af918b06aeb31268#egg=pytransfer
pytz==2018.5
pyviz-comms==0.7.0
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.3.2
qtconsole==4.2.0
QtPy==1.0
queuelib==1.5.0
rarfile==3.0
redis==2.10.3
requests==2.12.4
rope==0.9.4
sacred==0.7.4
scandir==1.9.0
scikit-learn==0.19.2
scipy==1.1.0
Scrapy==1.5.1
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.1
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.0.12
statsmodels==0.9.0
subprocess32==3.5.3
sympy==1.0
tensorboard==1.10.0
tensorboardX==1.2
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==0.9.0b1
toolz==0.9.0
torch==0.3.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
Twisted==18.7.0
unicodecsv==0.14.1
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wget==3.2
wrapt==1.10.11
xlrd==0.9.4
XlsxWriter==0.8.4
xlwt==1.0.0
zope.interface==4.5.0
